Refactor Analyzer's structure, and correct some false assumptions
This commit simplifies the code structure for the result-value checking logic in the CSharpFunctionalExtensions analyzer. It separates the checking mechanism into a new ResultValueWalker class to make it more maintainable.

This update also corrects the analyzer's handling of conditional statements and changes the messages to provide precise instructions to developers regarding the conditions to check before accessing a result object's value. This move will help prevent unexpected errors during runtime.

Additionally, the commit includes the creation of a WalkerResult class in a separate file, which stores the walking result. It also includes substantial changes in the UseResultValueWithoutCheckTests file, adding new cases to improve test coverage notably.
